+++ src/Schema/AbstractSchema.php
@@ -13,7 +13,9 @@ use Yiisoft\Db\Constraint\Constraint;
 use Yiisoft\Db\Constraint\IndexConstraint;
 use Yiisoft\Db\Exception\NotSupportedException;
 use Yiisoft\Db\Constant\GettypeResult;
+use Yiisoft\Db\Schema\Column\ColumnInterface;
 
+use function array_key_exists;
 use function gettype;
 use function is_array;
 
@@ -35,6 +37,11 @@ abstract class AbstractSchema implements SchemaInterface
      * @var string|null $defaultSchema The default schema name used for the current session.
      */
     protected string|null $defaultSchema = null;
+    /**
+     * @var (ColumnInterface|null)[] Saved columns from query results.
+     * @psalm-var array<string, ColumnInterface|null>
+     */
+    protected array $resultColumns = [];
     protected array $viewNames = [];
     private array $schemaNames = [];
     /** @psalm-var string[]|array */
@@ -58,6 +65,20 @@ abstract class AbstractSchema implements SchemaInterface
      * This allows {@see refresh()} to invalidate all cached table schemas.
      */
     abstract protected function getCacheTag(): string;
+
+    /**
+     * Returns the cache key for the column metadata received from the query result.
+     *
+     * @param array $metadata The column metadata from the query result.
+     */
+    abstract protected function getResultColumnCacheKey(array $metadata): string;
+
+    /**
+     * Creates a new column instance according to the column metadata received from the query result.
+     *
+     * @param array $metadata The column metadata from the query result.
+     */
+    abstract protected function loadResultColumn(array $metadata): ColumnInterface|null;
 
     /**
      * Loads all check constraints for the given table.
@@ -137,6 +158,39 @@ abstract class AbstractSchema implements SchemaInterface
             GettypeResult::NULL => DataType::NULL,
             default => DataType::STRING,
         };
+    }
+
+    final public function getResultColumn(array $metadata): ColumnInterface|null
+    {
+        if (empty($metadata)) {
+            return null;
+        }
+
+        $cacheKey = $this->getResultColumnCacheKey($metadata);
+
+        if (array_key_exists($cacheKey, $this->resultColumns)) {
+            return $this->resultColumns[$cacheKey];
+        }
+
+        $isCacheEnabled = $this->schemaCache->isEnabled();
+
+        if ($isCacheEnabled) {
+            /** @var ColumnInterface */
+            $this->resultColumns[$cacheKey] = $this->schemaCache->get($cacheKey);
+
+            if (isset($this->resultColumns[$cacheKey])) {
+                return $this->resultColumns[$cacheKey];
+            }
+        }
+
+        $column = $this->loadResultColumn($metadata);
+        $this->resultColumns[$cacheKey] = $column;
+
+        if ($column !== null && $isCacheEnabled) {
+            $this->schemaCache->set($cacheKey, $column, $this->getCacheTag());
+        }
+
+        return $column;
     }
 
     /**