+++ CHANGELOG.md
@@ -78,6 +78,7 @@
 - Enh #925, #951: Add callback to `Query::all()` and `Query::one()` methods (@Tigrov, @vjik)
 - New #954: Add `DbArrayHelper::arrange()` method (@Tigrov)
 - Chg #956: Remove nullable from `PdoConnectionInterface::getActivePdo()` result (@vjik)
+- New #945: Add option for typecasting values retrieved from DB (@Tigrov)
 - Enh #941: Add the ability for user-defined type casting (@Tigrov)
 - Enh #822: Refactor data readers (@Tigrov)